HHH-10873 - org.hibernate.type refactoring, created o.h.t.spi.StandardBasicTypes
