+++ hibernate-core/src/main/java/org/hibernate/dialect/Cache71Dialect.java
@@ -45,7 +45,7 @@ import org.hibernate.internal.util.StringHelper;
 import org.hibernate.persister.entity.Lockable;
 import org.hibernate.sql.CacheJoinFragment;
 import org.hibernate.sql.JoinFragment;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 
 /**
  * Cach&eacute; 2007.1 dialect.
