+++ hibernate-core/src/main/java/org/hibernate/dialect/HSQLDialect.java
@@ -48,7 +48,7 @@ import org.hibernate.internal.CoreMessageLogger;
 import org.hibernate.internal.util.JdbcExceptionHelper;
 import org.hibernate.internal.util.ReflectHelper;
 import org.hibernate.persister.entity.Lockable;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 
 import org.jboss.logging.Logger;
 import java.sql.DatabaseMetaData;
