+++ hibernate-core/src/main/java/org/hibernate/dialect/Ingres9Dialect.java
@@ -16,7 +16,7 @@ import org.hibernate.dialect.pagination.AbstractLimitHandler;
 import org.hibernate.dialect.pagination.LimitHandler;
 import org.hibernate.dialect.pagination.LimitHelper;
 import org.hibernate.engine.spi.RowSelection;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 
 /**
  * A SQL dialect for Ingres 9.3 and later versions.
