+++ hibernate-core/src/main/java/org/hibernate/dialect/Oracle9Dialect.java
@@ -20,7 +20,7 @@ import org.hibernate.hql.spi.id.global.GlobalTemporaryTableBulkIdStrategy;
 import org.hibernate.hql.spi.id.local.AfterUseAction;
 import org.hibernate.internal.CoreMessageLogger;
 import org.hibernate.internal.util.JdbcExceptionHelper;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 import org.jboss.logging.Logger;
 
 import java.sql.CallableStatement;
