+++ hibernate-core/src/main/java/org/hibernate/dialect/PostgresPlusDialect.java
@@ -13,7 +13,7 @@ import java.sql.Types;
 import org.hibernate.dialect.function.NoArgSQLFunction;
 import org.hibernate.dialect.function.NvlFunction;
 import org.hibernate.dialect.function.StandardSQLFunction;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 
 /**
  * An SQL dialect for Postgres Plus
