+++ hibernate-core/src/main/java/org/hibernate/dialect/function/StaticPrecisionFspTimestampFunction.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 import org.hibernate.QueryException;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
-import org.hibernate.type.StandardBasicTypes;
+import org.hibernate.type.spi.StandardBasicTypes;
 import org.hibernate.type.spi.Type;
 
 /**
